+++ AsBuiltReport.Veeam.VBR/Src/Private/Report/Get-AbrVbrDiagrammer.ps1
@@ -277,6 +277,8 @@ function Get-AbrVbrDiagrammer {
                             } else {
                                 Write-PScriboMessage -IsWarning "Unable to export the $DiagramType Diagram: '$FilePath' not found after generation."
                             }
+                        } else {
+                            Write-Host "Saved '$DiagramFilename' diagram to '$($OutputFolderPath)'." -InformationAction Continue
                         }
                     }
                 }